Share theme classification for native menu theme
